WIP: Implemented a factory pattern to create CrawlerService. Fixed english grammar for method isReachedTimeLimit in CrawlerThread class.
