+++ crawler-service/src/main/java/com/topcoder/productsearch/crawler/CrawlerThread.java
@@ -250,7 +250,7 @@ public class CrawlerThread implements Runnable {
    * @param url the url
    */
   private void enqueue(String url) {
-    if (crawlerService.getThreadPoolExecutor().isReachedTimeLimit(crawlerTask.getSite().getCrawlTimeLimit())) {
+    if (crawlerService.getThreadPoolExecutor().hasReachedTimeLimit(crawlerTask.getSite().getCrawlTimeLimit())) {
       logger.info("skip " + url + " , because of reached time limit");
       return;
     }
