+++ crawler-service/src/test/java/com/topcoder/productsearch/crawler/CrawlerThreadPoolExecutorTest.java
@@ -38,7 +38,7 @@ public class CrawlerThreadPoolExecutorTest {
 
     threadPoolExecutor.schedule(thread, 0, TimeUnit.MILLISECONDS);
     assertEquals(threadPoolExecutor.getRunningCount(), Integer.valueOf(1));
-    assertTrue(threadPoolExecutor.isReachedTimeLimit(0));
+    assertTrue(threadPoolExecutor.hasReachedTimeLimit(0));
     threadPoolExecutor.setExecutedHandler(runnable -> assertEquals(threadPoolExecutor.getRunningCount(),
         Integer.valueOf(0)));
     try {
